+++ source/B2B.Transactions.CimMessageAdapter.Tests/Stubs/TransactionIdsStub.cs
@@ -14,6 +14,7 @@
 
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using B2B.CimMessageAdapter.Messages;
 using B2B.CimMessageAdapter.Transactions;
 
 namespace MarketRoles.B2B.CimMessageAdapter.IntegrationTests.Stubs
