+++ source/B2B.Transactions.CimMessageAdapter/Messages/ITransactionIds.cs
@@ -14,7 +14,7 @@
 
 using System.Threading.Tasks;
 
-namespace B2B.CimMessageAdapter.Transactions
+namespace B2B.CimMessageAdapter.Messages
 {
     /// <summary>
     /// Store containing transaction id for all received market activity records
